backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.7.0
python-multipart==0.0.20
openai==1.57.0
stripe==11.3.0
google-auth==2.37.0
PyJWT==2.10.1
matplotlib==3.10.0
numpy==2.2.1
pypdf==5.1.0
Pillow==11.0.0
python-dotenv==1.0.1
email-validator==2.2.0
